Reset password content design
