+++ src/components/auth/contents/index.ts
@@ -1,4 +1,4 @@
 export { default as LoginContent } from './LoginContent'
 export { default as SignupContent } from './SignupContent'
-export { default as ForgotPasswordContent } from './ForgotPasswordContent'
+export { default as ResetPasswordContent } from './ResetPasswordContent'
 export { default as EnterLoginPasswordContent } from './EnterLoginPasswordContent'
